rename train -> train_sbd, train_multi -> train_sbd_multi
