+++ examples/fcis/train_sbd_multi.py
@@ -17,8 +17,8 @@ from chainercv.experimental.links import FCISTrainChain
 from chainercv.extensions import InstanceSegmentationVOCEvaluator
 from chainercv.links.model.ssd import GradientScaling
 
-from train import concat_examples
-from train import Transform
+from train_sbd import concat_examples
+from train_sbd import Transform
 
 
 def main():
